requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
semantic-kernel>=1.4.0
openai>=1.40.0
httpx>=0.27.0
aiohttp>=3.9.5
azure-identity>=1.16.0
